move client configurator
